Fix: move LNB directory into Layout
